Added support to mixinsSources. Print deprecate warning if define mixins in boot script.
